[Snippets] Added Manager::run_on_subgraph
