+++ src/common/snippets/include/snippets/pass/extract_constants.hpp
@@ -1,0 +1,29 @@
+// Copyright (C) 2018-2023 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#pragma once
+
+#include "subgraph_pass.hpp"
+
+namespace ov {
+namespace snippets {
+namespace pass {
+
+/**
+ * @interface ExtractConstants
+ * @brief Moves up Constants which aren't scalars from body to Subgraph and replace them with Parameters inside body
+ * @ingroup snippets
+ */
+class ExtractConstants: public SubgraphPass {
+public:
+    OPENVINO_RTTI("ExtractConstants", "0");
+    ExtractConstants() = default;
+
+    bool run_on_subgraph(const std::shared_ptr<op::Subgraph>& subgraph) override;
+};
+
+
+} // namespace pass
+} // namespace snippets
+} // namespace ov
